requirements.txt
pyairtable
python-dotenv
google-generativeai